表現の修正

+++ draft/convert.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
 for name in a-setup d-basic d-time d-vert ;do
-	wkhtmltopdf ${name}.html ${name}.pdf
-	rm ${name}.html
+	if [ -e ${name}.html ];then
+		wkhtmltopdf ${name}.html ${name}.pdf
+		rm ${name}.html
+	fi
 done
+
